\name{milano}
\alias{milano}
\docType{data}
\title{  Maximum annual rainfall series for different durations }
\description{
Maximum annual rainfall series for different durations recorded 
at the pluviograph of Palazzo Marino, Milan (Italy)  }
\usage{data(milano)}
\details{
Maximum annual precipitation series for 0.25, 0.5, 0.75, 1, 1.25, 1.50, 2, 2.5, 3, 4, 6 [h] 1931-1970 }
\source{
dataset of Palazzo Marino pluviograph , Milan (Italy)}
\examples{
data(milano)
## maybe str(series.rainfall) ; plot(series.rainfall) ...
}
\keyword{datasets}
