\name{Pistoia}
\alias{Pistoia}
\docType{data}
\title{ Dataset of Pistoia (Italy)
}
\description{ Bivariate series of observed rainfall-temperature for Pistoia (Italy) during the period 1951-2012
}
\usage{data(Pistoia)}
\format{
  A data frame with 744 observations on the following 2 variables.
  \describe{
    \item{\code{V1}}{Monthly cumulative rainfall (mm)}
    \item{\code{V2}}{Average monthly temperature (degree)}
  }
}

\source{ Ce.Spe.Vi. (Centro sperimentale per il vivaismo)
Web: http://www.cespevi.it
}

\examples{
data(Pistoia)
## maybe str(Pistoia) ; plot(Pistoia) ...
}
\keyword{datasets}
