\name{restrict}
\alias{restrict}
\alias{sync}
\alias{vec}
\alias{unvec}
\alias{modarma}
\alias{modfarma}
\alias{Stagiona}
\alias{arfimainf}
\alias{Arinf}
\alias{farima.err}
\alias{Mainf1}
\alias{Mainf3}
\alias{prova1}
\alias{prova2}
\alias{prova4}
\alias{prova7}
\alias{coeffd2}
\alias{filter.frac}
\alias{bd1}
\alias{Prodpoly1}
\alias{stlstu}
\alias{multides}
\alias{multipro}
\alias{multistag}
\alias{ClearRow}
\alias{id}
\alias{M}
\alias{permres}
\alias{PortBIC}
\alias{Portmanteau}
\alias{restrict}
\alias{sigmaZ}
\alias{simvar}
\alias{somdiag}
\alias{startvalue}
\alias{varcoeff}
\alias{Corrmatrix}
\alias{zero}
\alias{dippermres}
\alias{count}
\alias{Contazeri}
\alias{contaleva}
\alias{Levaneg}
\alias{aggiustam}
\alias{aggiusta}
\title{Support function}

\description{Support function}

\keyword{internal}
