\name{series.runoff}
\alias{series.runoff}
\docType{data}
\title{ Daily Runoff Series }
\description{
Daily runoff series of Tiber river observed to Ripetta station during the period 1930-1983 }
\usage{data(series.runoff)}

\details{29 february day values are removed to obtain integer periodical dataset 
  }
\source{
  Available on the web site www.gndci.cnr.it. "Gruppo nazionale per la difesa delle catastrofi idrogeologiche" 
}

\examples{
data(series.runoff)
## maybe str(series.runoff) ; plot(series.runoff) ...
}
\keyword{datasets}
