\name{Hourly.rainfall.series}
\alias{Hourly.rainfall.series}
\docType{data}
\title{  Hourly rainfall series }
\description{
Hourly rainfall series recorded in BUrlington (US) during the period 2012-2015.  }
\usage{data(Hourly.rainfall.series)}

\details{
Dataset is available on The Iowa Environmental Mesonet (IEM) website}
\source{
https://mesonet.agron.iastate.edu/request/download.phtml?}
\examples{
data(Hourly.rainfall.series)
## maybe str(series.rainfall) ; plot(series.rainfall) ...
}
\keyword{datasets}
