\name{LPM-package}
\alias{LPM-package}
\alias{LPM}
\docType{package}
\title{
\packageTitle{LPM}
}
\description{
\packageDescription{LPM}
}
\details{
See

\code{\link{ar.egls}},
\code{\link{lpm}}, 
\code{\link{mlpm}}
\code{\link{rain.adapt}}

}
\author{ 
Authors: Salvatore Grimaldi and Corrado Tallerini

Maintainer: \packageMaintainer{LPM}
}
\references{
Grimaldi S., Tallerini C., Serinaldi F. (2004)  'Modelli multivariati lineari per la generazione di serie di precipitazioni giornaliere' Giornata di Studio: Metodi Statistici e Matematici per l'Analisi Idrologiche  Napoli 2004

Grimaldi  S. , Serinaldi F. & Tallerini C. (2004)  'Multivariate linear parametric models applied to daily rainfall time series'  Mediterranean Storms, 6rd EGU Plinius Conference held in Mediterranean Sea, Italy, October 2004

Lutkepohl, H. (1993) Introduction to Multiple Time Series Analysis 2nd edition, Springer-Verlag, Berlin.

Grimaldi, S.,  'Linear parametric models applied on daily hydrological series', Journal of Hydrologic Engineering, Vol. 9, No 5 , September 2004.

Brockwell, P.J and Davis, R.A. (1990) Time Series: Theory and Methods 2nd edition, Springer, NY.

Hipel, K.W. and McLeod, A.I., (1994) Time Series Modelling of Water Resources and Enviromental Systems, Reading, UK.

     Hosking, J.R.M. (1980) 'The Multivariate Portmanteau Statistic' Journal of the American Statistical Association, Vol.75, N.371, 502-608.
}




